fix: adjust imports related to recent refactoring
