+++ engine/src/main/java/io/camunda/zeebe/process/test/engine/EngineStateMonitor.java
@@ -8,7 +8,7 @@
 package io.camunda.zeebe.process.test.engine;
 
 import io.camunda.zeebe.logstreams.storage.LogStorage;
-import io.camunda.zeebe.streamprocessor.StreamProcessor;
+import io.camunda.zeebe.stream.impl.StreamProcessor;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Timer;
